Admin auth resolved

+++ app/Http/Controllers/CategoriesController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Category;
+use Illuminate\Support\Facades\Auth;
 
 class CategoriesController extends Controller
 {
@@ -15,12 +16,19 @@ class CategoriesController extends Controller
     # display all categories
     public function index()
     {
+        if (Auth::id() == '18'){
+            return redirect('/admin');
+        }
+
         $categories = Category::all();
         return view('categories.index', ["categories" => $categories]);
     }
 
     public function create()
     {
+        if (Auth::id() != '18'){
+            return redirect('/categories');
+        }
         return view('categories.create');
     }
 
@@ -42,12 +50,20 @@ class CategoriesController extends Controller
 
     public function edit($id)
     {
+        if (Auth::id() != '18'){
+            return redirect('/categories');
+        }
+
         $category = Category::find($id);
         return view('category.edit', ["category" => $category]);
     }
 
     public function update($id)
     {
+        if (Auth::id() != '18'){
+            return redirect('/categories');
+        }
+
         $category = Category::find($id);
         $category->name = request('name');
         $category->save();
@@ -56,6 +72,9 @@ class CategoriesController extends Controller
 
     public function destroy($id)
     {
+        if (Auth::id() != '18'){
+            return redirect('/categories');
+        }
         $category = Category::find($id);
         $category->delete();
         return redirect()->route('category.index');
